bears/natural_language: Move reSTLintBear.py
Move `bears/natural_language/reSTLintBear.py`
to a new "bears/rest" folder.
The new file is now at 'bears/rest/reSTLintBear.py'

Move 'tests/natuarl_language/reSTLintBearTest.py'
to new "tests/rest" folder

Fixes https://github.com/coala-analyzer/coala-bears/issues/209
